backend/requirements.txt
# ========================================
# INTELLI-CREDIT Backend Dependencies
# ========================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# --- Database & Storage ---
supabase==2.7.0
postgrest==0.17.0

# --- Authentication ---
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0

# --- AI/ML ---
scikit-learn==1.5.0
xgboost==2.1.0
shap==0.46.0
numpy==1.26.0
pandas==2.2.0
joblib==1.4.0

# --- Agent Orchestration ---
langchain==0.3.0
langchain-core==0.3.0
langgraph==0.2.14
langchain-groq==0.2.0

# --- Document Processing ---
docling
PyMuPDF
easyocr

# --- Web Research ---
tavily-python==0.5.0

# --- Document Generation ---
python-docx==1.1.0
reportlab==4.2.0

# --- HTTP Client ---
httpx==0.27.0
aiohttp==3.10.0

# --- Utilities ---
python-dotenv==1.0.0
orjson==3.10.0
